Relocate public traits for the message encoding
